sequence counter as translation unit
