+++ belokurskaya.sofia/P1/sequence_counter.cpp
@@ -1,0 +1,34 @@
+#include "sequence_counter.hpp"
+#include <limits>
+#include <stdexcept>
+
+const int Min = std::numeric_limits<int>::min();
+
+SequenceCounter::SequenceCounter():
+  count_(0),
+  max_(Min),
+  count_max_(1)
+{}
+
+void SequenceCounter:: count(int number)
+{
+  size_t max_size = std::numeric_limits<size_t>::max();
+  if (count_ == max_size)
+  {
+      throw std::logic_error("sequence is too long");
+  }
+  if (number > max_)
+  {
+    max_ = number;
+    count_max_ = 1;
+  }
+  else if (number == max_)
+  {
+    ++count_max_;
+  }
+  ++count_;
+}
+size_t SequenceCounter::get_result() const
+{
+  return count_max_;
+}
